requirements.txt
Django==1.11.4
django-crispy-forms==1.6.1
django-multiselectfield==0.1.7
pytz==2017.2
